Cope with the CPSR differences between Apple's gdb and stock gdb.
The expectation is that you will use an alias or a bash function to help
determine whether the gdb about to be executed is Apple's version or
not.  If it is, then access CPSR as a structure (like the original
version).  If not, then treat CPSR as an integer value.